MinorChanges to everything

+++ SpriteClass.cpp
@@ -79,7 +79,7 @@ ID3D11ShaderResourceView* SpriteClass::GetTexture() {
 
 bool SpriteClass::InitializeBuffers(ID3D11Device* device) {
 	VertexType* vertices;
-	unsigned long* indices;
+	unsigned int* indices;
 	D3D11_BUFFER_DESC vertexBufferDesc, indexBufferDesc;
 	D3D11_SUBRESOURCE_DATA vertexData, indexData;
 	HRESULT result;
@@ -94,7 +94,7 @@ bool SpriteClass::InitializeBuffers(ID3D11Device* device) {
 
 	vertices = new VertexType[m_vertexCount];
 
-	indices = new unsigned long[m_indexCount];
+	indices = new unsigned int[m_indexCount];
 
 	memset(vertices, 0, (sizeof(VertexType) * m_vertexCount));
 
@@ -119,7 +119,7 @@ bool SpriteClass::InitializeBuffers(ID3D11Device* device) {
 	}
 
 	indexBufferDesc.Usage = D3D11_USAGE_DEFAULT;
-	indexBufferDesc.ByteWidth = sizeof(unsigned long) * m_indexCount;
+	indexBufferDesc.ByteWidth = sizeof(unsigned int) * m_indexCount;
 	indexBufferDesc.BindFlags = D3D11_BIND_INDEX_BUFFER;
 	indexBufferDesc.CPUAccessFlags = 0;
 	indexBufferDesc.MiscFlags = 0;
